+++ src/main/java/sixgaezzang/sidepeek/auth/dto/request/ReissueTokenRequest.java
@@ -1,6 +1,6 @@
 package sixgaezzang.sidepeek.auth.dto.request;
 
-import static sixgaezzang.sidepeek.auth.exeption.message.AuthErrorMessage.REFRESH_TOKEN_IS_NULL;
+import static sixgaezzang.sidepeek.auth.exception.message.AuthErrorMessage.REFRESH_TOKEN_IS_NULL;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
